… h2o-core/src/test/resources/hex/aucbuilder.csv.gz