@wendycwong this PR provides a significant speed-up of AUCBuilder in the histogram calculation. 

It is important to note that the original histogram bin merging was incorrect. The original code first calculates new histogram center of the merged bin and then uses the center to calculate delta error (it should instead use the original center of the histogram).

As a result of the fix the histograms will be slightly different as demonstrated in modified unit tests.

Performance measurement on Higgs dataset, 50 trees, 5 tree scoring interval:

- before: 417.861s, 436.055s, 413.959s, 417.398s, 436.362s
- after: 304.565s, 302.073s, 296.69s, 309.925s, 307.241s

This change brings roughly 25% overall speed-up for this dataset and settings.