@wendycwong 2017-12-18
Michal:

I believe I am still misleading you in my description of the AUC2.java.  Actually, I believe AUC2.java is not calculating the AUC but rather, it is trying to sample the ROC curve instead.  The AUC is the area under the ROC curve which is done in another step.  They insisted on using 400 points to sample the ROC curve.  Any time that number is exceeded, they will merge two points into one point using some kind of interpolation scheme calculated in compute_delta_error method.

If you look at the whole sample point merging processing, it is basically a O(n*(number of merges)).  For each merge, the mergeOneBin method takes O(n) because it needs to look at the possibility of merging each point and calculate the sqe and choose the point with the mininum sqe to merge.  Hence, the speedup here I believe is to reduce the operation per merge.

After looking at your PR, I do think there is errors in the sampling point merging operation.  I will go back and fix it.

Thank you for the analysis,
Wendy