@michalkurka 2017-12-19
@wendycwong you are right, the modified code is not directly calculating AUC. It is creating the histogram that is later used to calculate AUC (and also thresholds based on desired criteria we want to minimize, eg. precision and recall).

Please do _not_ include any of my changes in your branch. I would like to further work on this PR and I would like to avoid conflicts with your branch.

Regarding the bug I was referring to: please take a look at line 343 and 344. My fix was to swap their order. The previous code in the original order was calculating delta error based on the new histogram center (it should use the old center).