@michalkurka 2017-12-22
@wendycwong I found out that I was measuring the performance wrong: the actual speed-up is much better: 25% for the whole run instead of just 10%!