@michalkurka 2018-01-03
@h2o-ops please test!